# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CEPS Portal API — referral intake, authorizations, invoices, payments, remittances, vendors, reporting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors (preferred first)
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: w9Status
          in: query
          schema: { type: string }
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Vendor" }
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Add vendor (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Vendor detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404": { description: Not found }
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor (W-9 status, preferred flag, profile)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/w9:
    patch:
      operationId: uploadVendorW9
      tags: [vendors]
      summary: Attach an uploaded W-9 document (staff, or the vendor's own account)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorW9Input"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "403": { description: Forbidden }
        "404": { description: Not found }
  /vendors/{id}/contact:
    patch:
      operationId: updateVendorContact
      tags: [vendors]
      summary: Update vendor contact fields (staff, or the vendor's own account)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorContactInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "403": { description: Forbidden }
        "404": { description: Not found }
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, w9Status, preferred, active]
      properties:
        id: { type: string }
        name: { type: string }
        altaVendorNumber: { type: ["string", "null"] }
        ein: { type: ["string", "null"] }
        billingAddress: { type: ["string", "null"] }
        serviceAddress: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        w9Status: { type: string, enum: [pending, on_file, expired] }
        w9DocumentUrl: { type: ["string", "null"] }
        preferred: { type: boolean }
        active: { type: boolean }
        createdAt: { type: ["string", "null"] }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        altaVendorNumber: { type: string }
        ein: { type: string }
        billingAddress: { type: string }
        serviceAddress: { type: string }
        phone: { type: string }
        email: { type: string }
        contactPerson: { type: string }
        w9Status: { type: string, enum: [pending, on_file, expired] }
        preferred: { type: boolean }
    VendorUpdate:
      type: object
      properties:
        name: { type: string }
        altaVendorNumber: { type: string }
        ein: { type: string }
        billingAddress: { type: string }
        serviceAddress: { type: string }
        phone: { type: string }
        email: { type: string }
        contactPerson: { type: string }
        w9Status: { type: string, enum: [pending, on_file, expired] }
        w9DocumentUrl: { type: string }
        preferred: { type: boolean }
        active: { type: boolean }
    VendorW9Input:
      type: object
      required: [w9DocumentUrl]
      properties:
        w9DocumentUrl: { type: string, description: "Stored object path of the uploaded W-9 (e.g. /objects/uploads/uuid)" }
    VendorContactInput:
      type: object
      properties:
        email: { type: string }
        phone: { type: string }
        contactPerson: { type: string }
        billingAddress: { type: string }
        serviceAddress: { type: string }
...